refactor: register code objects tree with other trees
